Tweak powerhell setup

+++ template/setup.ps1
@@ -6,21 +6,20 @@ param (
 		[string]$MODCLASS
 )
 
-Remove-Item -Recurse -Force src/main/java/cam72cam/$MODID/Mod.java
+if(Test-Path -Path src/main/java/cam72cam/$MODID/Mod.java ){
+		Remove-Item -Recurse -Force src/main/java/cam72cam/$MODID/Mod.java
+}
 Remove-Item -Recurse -Force *gradle*
-
-New-Item -ItemType directory src/main/java/cam72cam/$MODID/
 
 Copy-Item -Path UniversalModCore/template/src/main/java/cam72cam/mod/Mod.java -Destination src/main/java/cam72cam/$MODID/
 Copy-Item -Path UniversalModCore/template/*gradle* -Destination $PWD -Recurse
 
-
 (Get-Content build.gradle) -replace "#MODID#", "$MODID" | Set-Content build.gradle
 (Get-Content build.gradle) -replace "#MODNAME#", "$MODNAME" | Set-Content build.gradle
 (Get-Content build.gradle) -replace "#MODVERSION#", "$MODVERSION" | Set-Content build.gradle
-(Get-Content src/main/java/cam72cam/mod/$MODID/Mod.java) -replace "#MODID#", "$MODID" | Set-Content src/main/java/cam72cam/mod/$MODID/Mod.java
-(Get-Content src/main/java/cam72cam/mod/$MODID/Mod.java) -replace "#MODNAME#", "$MODNAME" | Set-Content src/main/java/cam72cam/mod/$MODID/Mod.java
-(Get-Content src/main/java/cam72cam/mod/$MODID/Mod.java) -replace "#MODVERSION#", "$MODVERSION" | Set-Content src/main/java/cam72cam/mod/$MODID/Mod.java
+(Get-Content src/main/java/cam72cam/$MODID/Mod.java) -replace "#MODID#", "$MODID" | Set-Content src/main/java/cam72cam/$MODID/Mod.java
+(Get-Content src/main/java/cam72cam/$MODID/Mod.java) -replace "#MODNAME#", "$MODNAME" | Set-Content src/main/java/cam72cam/$MODID/Mod.java
+(Get-Content src/main/java/cam72cam/$MODID/Mod.java) -replace "#MODVERSION#", "$MODVERSION" | Set-Content src/main/java/cam72cam/$MODID/Mod.java
 
 Write-Host @"
 It's recommended you now run the following for intellij setup:
